requirements.txt
gunicorn==22.0.0
Flask>=3.0.3
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.5.1
joblib>=1.4.2
google-generativeai>=0.7.2
python-dotenv>=1.0.1

